+++ lizmap/forms/atlas_edition.py
@@ -1,6 +1,6 @@
 """Dialog for atlas edition."""
 
-from qgis.core import QgsMapLayerProxyModel, QgsProject
+from qgis.core import QgsMapLayerProxyModel
 
 from lizmap.definitions.atlas import AtlasDefinitions
 from lizmap.forms.base_edition_dialog import BaseEditionDialog
@@ -67,12 +67,6 @@ class AtlasEditionDialog(BaseEditionDialog, CLASS):
         if upstream:
             return upstream
 
-        wfs_layers_list = QgsProject.instance().readListEntry('WFSLayers', '')[0]
-        for wfs_layer in wfs_layers_list:
-            if layer.id() == wfs_layer:
-                break
-        else:
-            msg = tr(
-                'The layers you have chosen for this tool must be checked in the "WFS Capabilities"\n'
-                ' option of the QGIS Server tab in the "Project Properties" dialog.')
-            return msg
+        not_in_wfs = self.is_layer_in_wfs(layer)
+        if not_in_wfs:
+            return not_in_wfs