+++ lizmap/forms/time_manager_edition.py
@@ -123,15 +123,9 @@ class TimeManagerEditionDialog(BaseEditionDialog, CLASS):
             return upstream
 
         layer = self.layer.currentLayer()
-        wfs_layers_list = QgsProject.instance().readListEntry('WFSLayers', '')[0]
-        for wfs_layer in wfs_layers_list:
-            if layer.id() == wfs_layer:
-                break
-        else:
-            msg = tr(
-                'The layers you have chosen for this tool must be checked in the "WFS Capabilities"\n'
-                ' option of the QGIS Server tab in the "Project Properties" dialog.')
-            return msg
+        not_in_wfs = self.is_layer_in_wfs(layer)
+        if not_in_wfs:
+            return not_in_wfs
 
         if not self.start_field.currentField():
             return tr('Start attribute is mandatory.')