+++ lizmap/server/logger.py
@@ -3,43 +3,39 @@ __license__ = 'GPL version 3'
 __email__ = 'info@3liz.org'
 
 import functools
-import inspect
 import time
 import traceback
 
 from contextlib import contextmanager
-from pathlib import Path
 
 from qgis.core import Qgis, QgsMessageLog
 
 PLUGIN = 'Lizmap'
 
 
+# noinspection PyTypeChecker
 class Logger:
 
     @staticmethod
     def info(message: str):
-        QgsMessageLog.logMessage('{} : {}'.format(PLUGIN, message), PLUGIN, Qgis.Info)
+        QgsMessageLog.logMessage(str(message), PLUGIN, Qgis.Info)
 
     @staticmethod
     def warning(message: str):
-        QgsMessageLog.logMessage('{} : {}'.format(PLUGIN, message), PLUGIN, Qgis.Warning)
+        QgsMessageLog.logMessage(str(message), PLUGIN, Qgis.Warning)
 
     @staticmethod
     def critical(message: str):
-        QgsMessageLog.logMessage('{} : {}'.format(PLUGIN, message), PLUGIN, Qgis.Critical)
+        QgsMessageLog.logMessage(str(message), PLUGIN, Qgis.Critical)
 
     @staticmethod
     def log_exception(e: BaseException):
         """ Log a Python exception. """
-        QgsMessageLog.logMessage(
-            "Exception: {plugin}\n{e}\n{traceback}".format(
-                plugin=PLUGIN,
+        Logger.critical(
+            "Critical exception:\n{e}\n{traceback}".format(
                 e=e,
                 traceback=traceback.format_exc()
-            ),
-            PLUGIN,
-            Qgis.Critical
+            )
         )
 
 
@@ -66,7 +62,7 @@ def log_function(func):
     """ Decorator to log function. """
     @functools.wraps(func)
     def log_function_core(*args, **kwargs):
-        QgsMessageLog.logMessage('{}.{}'.format(PLUGIN, func.__name__), PLUGIN, Qgis.Info)
+        Logger.info(func.__name__)
         value = func(*args, **kwargs)
         return value
 
@@ -81,9 +77,7 @@ def profiling(func):
         start = time.time()
         result = func(*args, **kwargs)
         end = time.time()
-        Logger.info(
-            "{}.{}.{} ran in {}s".format(
-                PLUGIN, Path(inspect.stack()[1].filename).stem, func.__name__, round(end - start, 2)))
+        Logger.info("{} ran in {}s".format(func.__name__, round(end - start, 2)))
         return result
 
     return wrapper
@@ -95,9 +89,7 @@ def log_output_value(func):
     @functools.wraps(func)
     def wrapper(*args, **kwargs):
         result = func(*args, **kwargs)
-        Logger.info(
-            "{}.{}.{} output is {} for parameter {}".format(
-                PLUGIN, Path(inspect.stack()[1].filename).stem, func.__name__, result, str(args)))
+        Logger.info("{} output is {} for parameter {}".format(func.__name__, result, str(args)))
         return result
 
     return wrapper